# docs/generated/ranking_openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "Ranking Service",
    "version" : "0.0.1"
  },
...
  "components" : {
    "schemas" : {
      "ErrorResponse" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          }
        }
      },
...
      "PlayerRankingDTO" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "playerId" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "discordUsername" : {
            "type" : "string"
          },
          "lolAccountPPUID" : {
            "type" : "string"
          },
          "region" : {
            "$ref" : "#/components/schemas/Region"
          },
          "queueId" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "mmr" : {
            "type" : "integer",
            "format" : "int32"
          },
          "wins" : {
            "type" : "integer",
            "format" : "int32"
          },
          "losses" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
...
      "Region" : {
        "type" : "string",
        "enum" : [ "BR", "EUNE", "EUW", "LAN", "LAS", "NA", "OCE", "RU", "TR", "ME", "JP", "KR", "SEA", "TW", "VN", "PBE" ]
      },
...
      "UUID" : {
        "type" : "string",
        "format" : "uuid",
        "pattern" : "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}"
      },
...
    }
  },
  "paths" : {
...
    "/api/v1/ranking/players/{playerId}/queue-rankings" : {
      "get" : {
        "summary" : "Get all queue rankings for a player",
        "operationId" : "getPlayerRankings",
        "tags" : [ "Player Ranking" ],
        "parameters" : [ {
          "description" : "Player UUID",
          "example" : "550e8400-e29b-41d4-a716-446655440002",
          "name" : "playerId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully retrieved player ranking(s)",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PlayerRankingDTO"
                }
              }
            }
          },
          "404" : {
            "description" : "Player ranking not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ranking/players/{playerId}/queue-rankings/{queueId}" : {
      "get" : {
        "summary" : "Get a player's ranking for a specific queue",
        "operationId" : "getPlayerRankingByQueue",
        "tags" : [ "Player Ranking" ],
        "parameters" : [ {
          "name" : "playerId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        }, {
          "name" : "queueId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully retrieved player queue ranking",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PlayerRankingDTO"
                }
              }
            }
          },
          "404" : {
            "description" : "Player ranking not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers" : [ {
    "url" : "http://localhost:8080",
    "description" : "Auto generated value"
  }, {
    "url" : "http://0.0.0.0:8080",
    "description" : "Auto generated value"
  } ]
}